Refactor architecture
