+++ src/components/ClickeablePicture/ClickablePicture.js
