+++ src/components/Dice/Dice.js
