+++ src/components/IdCard/IdCard.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Avatar from './Avatar';
+import Avatar from '../Avatar/Avatar';
 
 const mapLabels = {
   firstName: 'First Name',
