+++ src/components/NumbersTable/NumbersTable.css
